refactor of analytical funcs
